Learn more about hsi
HSI has some funny design choices.

First, it does not understand absolute file paths.  So, the `cput`
command has to be invoked after an `lcd` and `cd` rather than just with
absolute paths.  This is insanity.  It also means that `cget` probably
needs to be fixed similarly.

ON THE OTHER HAND, `cput` is also smart enough to only update the tape
if the disk holds a newer file.  That makes life much easier.  No need
to manually do date comparisons etc.